feat: create magic link tests
